LicensePlate: Allow access to normalized value (even for invalid values)
For LicensePlates only the normalized value is now accessible to users
(via #toString), i.e. this is also done for LicensePlates that are
not valid.

This allows users of this class to continue working with the normalized
version of the LicensePlate, e.g. for searches in a database.

The class was also made immutable, which allows precomputing the normalized
value and the validation state.